Add support for multiple targets per server (#1)
This adds a load-balancer/fail-over feature that allows to specify multiple
backend targets per server. Targets will be attempted in random order.

This is a breaking change that requires a configuration update for existing
deployments.

All existing tests are still passing, and a new test was added to test the
new feature.