Make metrics server configurable
Add a block in the config file to allow the configuration of host, port and
mTLS details for the metrics server.

Luckily we can reuse a lot of the already existing logic in the code base,
but some code a new package named `tlsconn`.

The `readContent*` helpers were replaced by `os.ReadFile` from the
standard library.