Don't set default timeouts on listeners and targets
Currently, the config vaidator will set a timeout of 300 if the value isn't
specificed. So in other words, there is no way to set the timeout to 0,
which is a problem for long-standing connections.

This patch changes that, and leaves the default on 0 and makes sure the
deadlines are only set on connections that have a timeout configured.

`GenerateConfig` however will set this default, so the behaviour in
config-less scenario is retained.