Close connection to long-lived targets on proxy shutdown
When a proxy instance is requested to shut down, and the connection has no
timeout configured, we must close the connections forcefully so the calls
to `io.Copy()` return `EOF`.

For short-lived connections, we rely on the target connection to be closed
by the remote end.

Do this asynchronously so that we don't have to keep track of connections in
the proxy instance.

While at it, replace the `Quit` channel with a cancellabel context.