Close connection on proxy shutdown
When a proxy instance is requested to shut down, we must close the
connections individually so the calls to `io.Copy()` return `EOF`.

Do this asynchronously so that we don't have to keep track of connections in
the proxy instance.

While at it, replace the `Quit` channel with a cancellabel context.